Split event and metric tests
